Gets rid of the split-screen that appears soon after startup.
Accomplished by removing the map in the righthand side of the split-screen.

Had to do this because while using the GTK look-and-feel, it seems to be
impossible to get rid of the split-screen.

The small buttons which control the split-screen apparently don't exist
in the GTK look-and-feel.  (In the usual look-and-feel, the buttons are
near the top of the split-screen's divider.)

The "Split Screen" menu item in the "View" menu also doesn't do
anything.

Nor does "Ctrl+Shift+Back Slash", the keyboard shortcut for that
menu item.

Would fix this if I knew how.