Affects VUE on Unix.
Changed how URLs and filepaths are handled when you click on
the resource buttons of bubbles with attached URL or file
resources.

Instead of trying to open files with GNOME or KDE, VUE runs
this shell command:

"for_vue---filepath_opener"

...using the filepath as an argument.

On my system, that shell command is simply a symlink to
Puppy Linux's usual default file manager, Rox-Filer.

And, instead of trying to open web URLs in Firefox or
Netscape(!), VUE runs this shell command:

"for_vue---weburl_opener"

...using the URL as an argument.

On my system, that shell command is a symlink to the
Pale Moon web browser.